Float sine example and anyTouch init fix
In examples/Fundamentals/sine.cpp, switched to sinf(); in ctrlInputs.cpp
fixed initialization error of anyTouch (set to 0 instead if -1).

+++ examples/Fundamentals/sine/render.cpp
@@ -42,7 +42,7 @@ void render(LDSPcontext *context, void *userData)
 	{
 		float out = amplitude * sinf(phase);
 		phase += 2.0f * (float)M_PI * frequency * inverseSampleRate;
-		while(phase > 2.0f *M_PI)
+		while(phase > 2.0f * (float)M_PI)
 			phase -= 2.0f * (float)M_PI;
 		
 		for(int chn=0; chn<context->audioOutChannels; chn++)
